ジュースのクラスを作成し在庫数確認の機能を追加
