+++ ruby/vending_machine/main.rb
@@ -1,4 +1,5 @@
-require_relative 'vending_machine'
+require_relative 'suica'
+require_relative 'juice'
 
 suica = Suica.new
 
@@ -9,3 +10,7 @@ p suica.check_charge_amount
 p suica.charge(200)
 
 p suica.check_charge_amount
+
+juice = Juice.new(:pepsi)
+
+p juice.check_stock
